Move factored-out functions to separate crate.
