+++ Cargo.toml
@@ -1,6 +1,7 @@
 [workspace]
 
 members = [
+    "mijia",
     "publish-mqtt",
     "read-all-devices",
 ]
