+++ mijia/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "mijia"
+version = "0.1.0"
+authors = ["Luis Félix <lcs.felix@gmail.com>", "Andrew Walbran <qwandor@google.com>"]
+edition = "2018"
+license = "MIT OR Apache-2.0"
+
+[dependencies]
+blurz = "0.4.0"
+regex = "1"
+lazy_static = "1.4.0"
