+++ read-all-devices/Cargo.toml
@@ -7,5 +7,6 @@ license = "MIT OR Apache-2.0"
 
 [dependencies]
 blurz = "0.4.0"
-regex = "1"
 lazy_static = "1.4.0"
+mijia = { path = "../mijia" }
+regex = "1"
